+++ android/app/src/main/AndroidManifest.xml
@@ -65,10 +65,18 @@
         </service>
       <!-- For background messaging -->
     <service android:name="io.invertase.firebase.messaging.RNFirebaseBackgroundMessagingService" />
-    <meta-data
+    <!-- <meta-data
     android:name="com.google.firebase.messaging.default_notification_channel_id"
-    android:value="@string/default_notification_channel_id" />
-
+    android:value="@string/default_notification_channel_id" /> -->
+    <receiver android:name="io.invertase.firebase.notifications.RNFirebaseNotificationReceiver"/>
+    <receiver android:enabled="true" android:exported="true" android:name="io.invertase.firebase.notifications.RNFirebaseNotificationsRebootReceiver">
+      <intent-filter>
+        <action android:name="android.intent.action.BOOT_COMPLETED"/>
+        <action android:name="android.intent.action.QUICKBOOT_POWERON"/>
+        <action android:name="com.htc.intent.action.QUICKBOOT_POWERON"/>
+        <category android:name="android.intent.category.DEFAULT" />
+      </intent-filter>
+    </receiver>
   </application>
 
 </manifest>
